+++ lib/flop/relay.ex
@@ -55,7 +55,8 @@ defmodule Flop.Relay do
   ## Options
 
   - `:cursor_value_func`: 2-arity function that takes an item from the query
-    result and the `order_by` fields and returns the unencoded cursor value.
+    result and the cursor fields and returns the unencoded cursor value. The
+    cursor fields are the order fields plus the tiebreaker.
   """
   @doc since: "0.8.0"
   @spec connection_from_result({[any], Meta.t()}, [Flop.option()]) ::
@@ -165,7 +166,8 @@ defmodule Flop.Relay do
   ## Options
 
   - `:cursor_value_func`: 2-arity function that takes an item from the query
-    result and the `order_by` fields and returns the unencoded cursor value.
+    result and the cursor fields and returns the unencoded cursor value. The
+    cursor fields are the order fields plus the tiebreaker.
   """
   @doc since: "0.8.0"
   @spec edges_from_result({[{any, any}] | [any], Meta.t()}, [Flop.option()]) ::